% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdbscan.R
\name{dmdbscan}
\alias{dmdbscan}
\title{Dynamic Method DBSCAN (DMDBSCAN) Function}
\usage{
dmdbscan(
  data,
  MinPts,
  scale = FALSE,
  method = c("hybrid", "raw", "dist"),
  seeds = TRUE,
  showplot = FALSE,
  countmode = NULL
)
}
\arguments{
\item{data}{data matrix, data.frame, dissimilarity matrix or 'dist'-object. Specify 'method="dist"' if the data should be interpreted as dissimilarity matrix or object. Otherwise Euclidean distances will be used.}

\item{MinPts}{Reachability minimum no. of points, see Ester et al. (1996).}

\item{scale}{scale the data if 'TRUE'.}

\item{method}{"dist" treats data as distance matrix (relatively fast but memory expensive), "raw" treats data as raw data and avoids calculating a distance matrix (saves memory but may be slow), "hybrid" expects also raw data, but calculates partial distance matrices (very fast with moderate memory requirements).}

\item{seeds}{FALSE to not include the isseed-vector in the dbscan-object.}

\item{showplot}{0 = no plot, 1 = plot per iteration, 2 = plot per subiteration.}

\item{countmode}{NULL or vector of point numbers at which to report progress.}
}
\value{

}
\description{
Dynamic Method DBSCAN (DMDBSCAN) Function
}
